#Categories [UPDATE] updated json

+++ pages/api/categories/[id].js
@@ -33,10 +33,10 @@ const updateCategory = async (req, res) => {
 
     res.json({
       msg: "Update Success!",
-      // category: {
-      //   ...newCategory._doc,
-      //   name,
-      // },
+      category: {
+        ...newCategory._doc,
+        name,
+      },
     });
   } catch (err) {
     return res.status(500).json({ err: err.message });
@@ -51,12 +51,12 @@ const deleteCategory = async (req, res) => {
 
     const { id } = req.query;
 
-    // const products = await Products.findOne({ category: id });
+    const products = await Products.findOne({ category: id });
 
-    // if (products)
-    //   return res.status(400).json({
-    //     err: "Please delete all products with a relationship",
-    //   });
+    if (products)
+      return res.status(400).json({
+        err: "Please delete all products with a relationship",
+      });
 
     await Categories.findByIdAndDelete(id);
